ConcatenatedIterator now has its own child stack
We're creating a lot of recursive ConcatenatedIterator instances, which means that
every .next() on a top-level instance calls deep into an implicit tree of instances.
This .next() chain can eventually get out of hand, so ConcatenatedIterator now
maintains its own essential call stack. The top-level instance always peeks at the
top of its stack to determine which level of child it should call .next() on,
eliding the rest of the .next() call chain.