Replace PersistentSet with IdentitySet
The goal behind this change was to handle situations where two distinct objects
with the same hashCode are added to the same set. PersistentSet internally just
goes off of hashCode, but IdentitySet also ensures that different identities
can "live" in the same bucket.

This change also gave me the opportunity to pit QPTrie against the internals
of kotlinx.collections.immutable.*... and QPTrie definitely didn't win, so
there's some optimizations to be made still.