Internal radix 64 encoding scheme
These are Radix 64 encoding scheme that can
losslessly and reversibly translate arbitrary
lists of either just strings at a low level,
or valid JSON definitions of "null",
"boolean", "number", and "string" at a high
level. The resulting encoding preserves sort
order within its byte representation, which
will become necessary for range queries.

JSON is a hierarchical key-value definition,
but we can't internally work with
hierarchies. What we can do instead is
realize the key hierarchy as a flat "key
path" within the low-level encoding, and then
encode the values with the higher level
encoding. We avoid the higher-level encoding
for the "key path" because keys in JSON are
always strings, and don't need additional
type information.