Also report on CloudEvent non-data attributes
The output into queries now takes the form of Structured Content Mode
as defined in the HTTP Protocol Binding for CloudEvents, Version 1.0.3-wip.
That is to say, the result will be a JSON object with "source", "id", and
"data" fields, where the "data" is the contents previously reported to
clients directly.

The query specification still works over "data", unless a special prefix
"../" is used instead of "/" in JSON Pointer path encoding. This tells
Thorium to start its query from the root of the document, allowing clients
to filter based on CloudEvent attribute metadata instead.