Correctly handle starts-with for not-3-divisible value lengths
Previously, a query of the form
   key=~value

would fail, because in this literal case "value" has 5 characters,
which is not divisible by 3, resulting in trailing empty bits
that would not be properly matched due to the Radix64 conversion.
For example, this should have matched:

   { "key": "valuee" }

but previously would not.

The fix for this issue is to internally perform a starts-with query
over all but the final byte for any encoded value with a length
not divisible by 4 (derived from the 3-to-4 byte expansion of the
Radix64 encoding process), and establish an additional post-matching

   data[key][offset] & mask == expected

check within the FullQuery methods.