Native transports, and always depend on native macOS DNS resolvers
Previously, we'd only install the macOS DNS resolver dependency for
Netty if we were running on macOS, and even then we'd only depend on
the version for our current running architecture. That was... fine...
for prototyping but for actual delivery of a single artifact, we'll
always have to depend on them. Thankfully, having inert "wrong target"
libraries in a JAR only increases the artifact size, and doesn't
otherwise cause problems.