Make Linux native-image work by runtime-initializing KQueue
The native-image builds attempt to load the KQueue libraries...
for some reason. This causes Linux builds to fail. However, by
shifting the initialization of the KQueue class to runtime instead
of build-time, we can build properly on Linux.

Interestingly, native-image on macOS has no problem with the inert
EPoll libraries.

This change doesn't prevent the native image from correctly running
on either platform, so it's safe to include in general.