Make tail call optimization impossible in KeyValueSizeLimits
Tail call optimization isn't something implemented by the JVM, but
there's a possibility that a GraalVM optimization for native images
might attempt it. To ensure KeyValueSizeLimits always works, we'll
add 1 to the tested number on each call -- this will still be
ignored later, because the callback returns the actual tested call,
is always lower than the return from recurseUntilZero, and is
chosen as the "maximum key size" over recurseUntilZero because
it is lower.